pyproject.toml
[project]
name = "starlight-scribe"
version = "0.1.0"
description = "Transform rough parenting notes into illustrated children's books"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "google-genai>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "weasyprint>=62.0",
    "Pillow>=10.0.0",
    "Jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
starlight = "main:cli"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
